knowledgebase/serializers.py
from rest_framework import serializers
from .models import FAQ, BookingDetail

class FAQSerializer(serializers.ModelSerializer):
    class Meta:
        model = FAQ
        fields = ['question', 'answer']

class BookingDetailSerializer(serializers.ModelSerializer):
    class Meta:
        model = BookingDetail
        fields = ['contact', 'location']
