requirements.txt
Django>=4.0
djangorestframework>=3.14
gunicorn>=20.1.0
gspread>=5.10.0
oauth2client>=4.1.3
pandas>=2.2.0
openpyxl>=3.1.2
